Connected postgres DB to server and front end; limit 100 items returned for now
