+++ package.json
@@ -25,6 +25,7 @@
     "lodash-webpack-plugin": "^0.11.5",
     "mysql": "^2.17.1",
     "nodemon": "^1.19.1",
+    "pg": "^7.12.0",
     "react": "^16.8.6",
     "react-bootstrap": "^1.0.0-beta.9",
     "react-bootstrap-typeahead": "^3.4.6",
